requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0

# PDF generation
pdfkit>=1.0.0

# Template rendering
jinja2>=3.1.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0

# HTTP requests (for ACM API)
requests>=2.31.0

# Optional: Better formatting
Pillow>=10.0.0